Add additional logic for included files in package
Homebrew can't reuse a Keg exported from this tool because of the way
the rsync copies the files and dereferences the symlinks. This still
produces a working package, but Homebrew complains about it. This commit
adds in the additional logic needed to produce a tree in the package
that let's Homebrew reference and use the Keg as if Homebrew had
installed the Keg rather than via a pkg.